src/app/requirements.txt
# Core AI dependencies
google-generativeai>=0.3.0

# FastAPI framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# For async file operations
aiofiles>=23.0.0

# For image processing (if needed)
pillow>=10.0.0

# Environment variables
python-dotenv>=1.0.0

# Additional utilities
python-multipart>=0.0.6  # For file uploads